+++ example/lib/src/feature/home/cubit/home_user_with_delay_cubit.dart
@@ -1,7 +1,7 @@
 import 'package:dio_nexus/dio_nexus.dart';
+import 'package:example/src/feature/home/service/IHomeService.dart';
+import 'package:example/src/feature/home/service/home_service.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:network_manager_test/src/feature/home/service/IHomeService.dart';
-import 'package:network_manager_test/src/feature/home/service/home_service.dart';
 
 import '../model/users_model.dart';
 
@@ -17,8 +17,8 @@ class HomeUserDelayCubit extends Cubit<ResultState<Users>> {
     IResponseModel<Users?>? response = await homeService.getUsersWithDelay(5);
     if (response?.errorModel?.networkException != null) {
       emit(ResultState.error(error: response!.errorModel!.networkException!));
-    } else if (response?.model != null) {
-      emit(ResultState.data(data: response!.model!));
+    } else if (response?.data != null) {
+      emit(ResultState.data(data: response!.data!));
     } else {
       emit(const ResultState.idle());
     }