+++ example/lib/src/feature/home/cubit/home_users_cubit.dart
@@ -1,6 +1,6 @@
 import 'package:dio_nexus/dio_nexus.dart';
-import 'package:network_manager_test/src/feature/home/service/IHomeService.dart';
-import 'package:network_manager_test/src/feature/home/service/home_service.dart';
+import 'package:example/src/feature/home/service/IHomeService.dart';
+import 'package:example/src/feature/home/service/home_service.dart';
 import '../model/users_model.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -18,8 +18,8 @@ class HomeUsersCubit extends Cubit<ResultState<Users>> {
     IResponseModel<Users?>? response = await homeService.getUsers();
     if (response?.errorModel?.networkException != null) {
       emit(ResultState.error(error: response!.errorModel!.networkException!));
-    } else if (response?.model != null) {
-      emit(ResultState.data(data: response!.model!));
+    } else if (response?.data != null) {
+      emit(ResultState.data(data: response!.data!));
     } else {
       emit(const ResultState.idle());
     }