+++ lib/src/constants/variable_constants.dart
@@ -1,0 +1,11 @@
+/// The VariableConstants class is a singleton class that contains constants for the application.
+class VariableConstants {
+  /// The constructor of the VariableConstants class.
+  VariableConstants._();
+
+  /// The time for the toast.
+  static const int toastDuration = 3;
+
+  /// The font size of the toast.
+  static const double toastFontSize = 16;
+}