+++ lib/src/interface/IErrorModel.dart
@@ -1,15 +1,22 @@
-// ignore_for_file: public_member_api_docs, sort_constructors_first
-import 'package:dio_nexus/dio_nexus.dart';
+import 'package:dio_nexus/src/network/index.dart';
 
+/// The IErrorModel interface is a generic interface that defines the structure of an error model.
 abstract class IErrorModel {
-  int? statusCode;
-  String? errorMessage;
-  NetworkExceptions? networkException;
+  /// The constructor of the IErrorModel class.
   IErrorModel(
     this.statusCode,
     this.errorMessage,
     this.networkException,
   );
+
+  /// The status code of the error.
+  int? statusCode;
+
+  /// The error message of the error.
+  String? errorMessage;
+
+  /// The network exception of the error.
+  NetworkExceptions? networkException;
 
   @override
   String toString() =>