+++ lib/src/interface/index.dart
@@ -1,3 +1,4 @@
-export './IDioNexusManager.dart';
-export './IDioNexusNetworkModel.dart';
-export './IResponseModel.dart';
+export 'IDioNexusManager.dart';
+export 'IDioNexusNetworkModel.dart';
+export 'IErrorModel.dart';
+export 'IResponseModel.dart';