+++ lib/src/languages/language_tr.dart
@@ -1,67 +1,78 @@
-import 'abstract/languages.dart';
+import 'package:dio_nexus/src/languages/index.dart';
 
+/// The `LanguageTr` class extends the `Languages` abstract class and provides the Turkish language translations.
 class LanguageTr extends Languages {
   @override
-  String get networkConnectionNoInternetConnection => "İnternet bağlantısı yok";
+  String get networkConnectionNoInternetConnection => 'İnternet bağlantısı yok';
 
   @override
-  String get networkConnectionTryAgain => "Tekrar Dene";
+  String get networkConnectionTryAgain => 'Tekrar Dene';
 
   @override
-  String get notImplemented => "Uygulanmadı";
+  String get notImplemented => 'Uygulanmadı';
 
   @override
-  String get requestCancelled => "İstek İptal Edildi";
+  String get requestCancelled => 'İstek iptal edildi';
 
   @override
-  String get internalServerError => "Dahi̇li̇ sunucu hatası";
+  String get internalServerError => 'Dahi̇li̇ sunucu hatası';
 
   @override
-  String get notFound => "Bulunamadı!";
+  String get notFound => 'Bulunamadı!';
 
   @override
-  String get serviceUnavailable => "Servis kullanılamıyor";
+  String get serviceUnavailable => 'Servis kullanılamıyor';
 
   @override
-  String get methodNotAllowed => "Yönteme izin verilmiyor";
+  String get methodNotAllowed => 'Yönteme izin verilmiyor';
 
   @override
-  String get badRequest => "Geçersiz istek";
+  String get badRequest => 'Geçersiz istek';
 
   @override
-  String get unauthorisedRequest => "Yetkisiz istek";
+  String get unauthorizedRequest => 'Yetkisiz istek';
 
   @override
-  String get unexpectedError => "Beklenmedik hata oluştu";
+  String get unexpectedError => 'Beklenmedik hata oluştu';
 
   @override
   String get requestTimeout =>
-      "Bağlantı başarısız oldu. Zayıf internet bağlantısı";
+      'Bağlantı başarısız oldu. Zayıf internet bağlantısı';
 
   @override
-  String get conflict => "Bir çakışma nedeniyle hata oluştu";
+  String get conflict => 'Bir çakışma nedeniyle hata oluştu';
 
   @override
-  String get sendTimeout => "Veri Gönderme Süresi Aşımı";
-  
+  String get sendTimeout => 'Veri gönderme süresi aşımı';
+
   @override
-  String get unableToProcess => "Veriler işlenemiyor";
-  
+  String get unableToProcess => 'Veriler işlenemiyor';
+
   @override
-  String get formatException => "Beklenmedik hata oluştu";
-  
+  String get formatException => 'Beklenmedik hata oluştu';
+
   @override
-  String get notAcceptable => "Kabul edilemez";
-  
+  String get notAcceptable => 'Kabul edilemez';
+
   @override
-  String get badCertificate => "Geçersiz veya Güvenilmeyen SSL/TLS sertifikası";
-  
+  String get badCertificate => 'Geçersiz, güvenilmeyen SSL/TLS sertifikası';
+
   @override
-  String get connectionError => "İstek için bağlantı kurulamadı";
-  
+  String get connectionError => 'İstek için bağlantı kurulamadı';
+
   @override
-  String get receiveTimeout => "Sunucu yanıt vermiyor. Lütfen tekrar deneyiniz";
-  
+  String get receiveTimeout => 'Sunucu yanıt vermiyor. Lütfen tekrar deneyiniz';
+
   @override
-  String get forbiddenRequest => "Yasak İstek";
+  String get forbiddenRequest => 'Yasak istek';
+
+  @override
+  String get rateLimitExceeded => 'İstek limiti aşıldı';
+
+  @override
+  String get badGateway => 'Hatalı Ağ Geçidi: Sunucu geçersiz yanıt aldı';
+
+  @override
+  String get gatewayTimeout =>
+      'Ağ Geçidi Zaman Aşımı: Sunucu zamanında yanıt veremedi';
 }