+++ lib/src/network/network_model_parser.dart
@@ -1,5 +1,75 @@
 part of '../dio_nexus_manager.dart';
 
+/// [R] is the type of the response data.
+/// Example:
+/// [R] can be Model or [List<Model>]
+///
+/// [T] is the type of the response model.
+/// [T] must be Model
+/// [T] has fromJson method and fromJson method
+R? _parseResponseData<T extends IDioNexusNetworkModel<T>, R>(
+  T responseModel,
+  dynamic data, {
+  bool? printLogsDebugMode,
+}) {
+  try {
+    if (data is List) {
+      return data
+          .map((e) => responseModel.fromJson(e as Map<String, dynamic>))
+          .toList() as R;
+    } else if (data is Map<String, dynamic>) {
+      return responseModel.fromJson(data) as R;
+    } else {
+      CustomLogger(
+        data: '$data data is [${data.runtimeType}] cannot be parsed as $R',
+      ).show(printLog: printLogsDebugMode ?? false);
+    }
+  } catch (err) {
+    CustomLogger(data: " $err \n\n $R CAN'T PARSE TO $data")
+        .show(printLog: printLogsDebugMode ?? false);
+  }
+  return null;
+}
+
+/// R is primitive type
+/// Example:
+/// [R] can be String, int, double, bool
+/// [R] can be List<String>, List<int>, List<double>, List<bool>
+R? _parsePrimitiveResponseData<R>(dynamic data, {bool? printLogsDebugMode}) {
+  try {
+    if (data is String || data is int || data is double || data is bool) {
+      return data as R;
+    } else if (data is List) {
+      return _parsePrimitiveListData<R>(data);
+    } else {
+      CustomLogger(
+        data: '$data primitive data cannot be parsed as $R',
+      ).show(printLog: printLogsDebugMode ?? false);
+    }
+  } catch (err) {
+    CustomLogger(data: " $err \n\n $R CAN'T PARSE TO $data")
+        .show(printLog: printLogsDebugMode ?? false);
+  }
+  return null;
+}
+
+/// List<R> is the type of the response data.
+/// Example:
+/// [R] can be String, int, double, bool
+/// [R] can be List<String>, List<int>, List<double>, List<bool>
+R? _parsePrimitiveListData<R>(List<dynamic> datas) {
+  if (datas.every((e) => e is String)) {
+    return List<String>.from(datas) as R;
+  } else if (datas.every((e) => e is int)) {
+    return List<int>.from(datas) as R;
+  } else if (datas.every((e) => e is double)) {
+    return List<double>.from(datas) as R;
+  } else if (datas.every((e) => e is bool)) {
+    return List<bool>.from(datas) as R;
+  }
+  return null;
+}
+
 /// Must be top-level function
 dynamic _parseAndDecode(String response) {
   return jsonDecode(response);
@@ -16,52 +86,5 @@ dynamic parseJson(String body) async {
     return await compute(_parseAndDecode, body);
   } catch (e) {
     return body;
-  }
-}
-
-R? _modelResponseData<T extends IDioNexusNetworkModel<T>, R>(
-  T responseModel,
-  dynamic responseData,
-  bool? printLogsDebugMode,
-) {
-  try {
-    if (responseData is List) {
-      return responseData
-          .map((e) => responseModel.fromJson(e as Map<String, dynamic>))
-          .toList() as R;
-    } else if (responseData is Map<String, dynamic>) {
-      return responseModel.fromJson(responseData) as R;
-    } else if ((responseModel is NexusModel) &&
-        (responseModel as NexusModel).modelType == responseData.runtimeType) {
-      return genericNexusModelConverter(responseData) as R;
-    } else {
-      CustomLogger(
-        data:
-            '$responseData data is [${responseData.runtimeType}] cannot be parsed as $R',
-      ).show(printLogsDebugMode ?? false);
-    }
-  } catch (err) {
-    CustomLogger(data: " $err \n\n $R CAN'T PARSE TO $responseData")
-        .show(printLogsDebugMode ?? false);
-  }
-  return null;
-}
-
-/// The function converts a dynamic data object into a generic Nexus model.
-///
-/// Args:
-///   data (dynamic): The "data" parameter is a dynamic variable, which means it can hold any type of
-/// data. It is used as input to the method "genericNexusModelConverter".
-dynamic genericNexusModelConverter(dynamic data) {
-  if (data.runtimeType == int) {
-    return NexusModel<int>(value: data as int);
-  } else if (data.runtimeType == String) {
-    return NexusModel<String>(value: data as String);
-  } else if (data.runtimeType == bool) {
-    return NexusModel<bool>(value: data as bool);
-  } else if (data.runtimeType == double) {
-    return NexusModel<double>(value: data as double);
-  } else if (data.runtimeType == dynamic) {
-    return NexusModel<dynamic>(value: data);
   }
 }